requirements.txt
ruamel.yaml==0.16.5
fabric==2.5.0
beautifultable==1.0.0